Remove hack; address issue w/ 5.7
The "hack" solution looks like it was addressing a problem in Core,
which has been corrected in WordPress 5.7. However, there is still an
issue in 5.7, but it's specifically with the positioning of the button
we add for link selection in the url-picker.

This commit does two things:

- remove hacky solution
- fix issue w/ url-picker in 5.7

In < 5.7 the original issue will persist, but I feel that's a reasonable
tradeoff since

1. The site will probably be updated in short order to 5.7 (if it hasn't
   already by launch)
2. The problem is a minor visual glitch; not a feature issue.

As for why I don't want to just leave the "hack" in: It's messing with
styling on deep core-generated classes, so the side effects of the
modification are hard to predict. I'd rather remove that potential
source of bugs with much for targeted, controllable CSS.