Make the whole of the card clickable
The old site had a solution where a link was 'put over' the contents
of the card artificially. I like this much better, because clicking
the whole card is always meant for mouse users. Keyboard users can
simply use the link. As an added bonus this makes the cards more
predictable for screen readers.