Use file hash as version string for theme style and script files
Ensures that front-end styles are cache-busted on any updates to the
stylesheet or script files.

I don't particularly like this approach, but I couldn't think of any way
of caching this hash that wouldn't risk breaking on a deploy if someone
didn't manually purge the object cache, and VIP doesn't have an API for
automatically purging this on deploys.