Remove extra content in additional fields setting
The additional fields setting was intended to be used for hidden
inputs. I conclude this from the name and the custom kses handling.
But, on production it has extra content. In Gutenberg, we allow this
content to be set in the editor. Making the use of this extra content
unnecessary. So, simply strip the content that is currently there on
production. As soon as all pages are Gutenberg, this content can be
removed from the setting and this line of code can be removed.