Build changes from 30d1aac26e91861b84fa410208502094e5752281
Make link-column pattern dynamic

Because of the nature of the block, this pattern included a hardcoded
url (the location of the svg spritesheet). This url would then differ
between environments, causing errors when trying to use the pattern on
preprod.

This modified the pattern so that it dynamically generates the correct
link in the environment in which it finds itself, (hopefully)
resolving the issue.