Implement double heading block
The first-insert flow is as follows:

- The primary heading acts as a standard RichText.
- The secondary heading acts as a standard RichText until you add some
content.
- From then on there is a button that shows translated headings.
- When at least one translated heading is present this is shown with a
a square-bracket placeholder in the non-expanded view.

This follows the same design as the home page hero.

The way I've set this up means that this block can be copy and pasted
from one language to another very easily. It should automatically
adjust the frontend view based on the site language.