Resolve issue w/ "blinking backdrop"
This reverts a change that attempted to make the behavior of the
dropdown a little more "consistent" but which introduced other, greater,
problems. Ultimately the "state" of the dropdown (as defined by
`data-open`) is not necessarily relevant to the client user: It's a tool
to manage visual state *if needed*. In the case of this menu, it isn't
needed on desktop and, crucially, the aria attributes it manages are
also not necessary (because they apply only to the toggle, which is not
visible on desktop). All the matters on desktop is the presence (or
rather, absence) of the `hidden` attribute on the drawer, which is a
*side-effect* of the dropdown state, not a *determiner* of it. Therefore
I feel fairly comfortable manipulating that property "outside" of the
behavior of the dropdown.

Future issues like this should be taken on a case-by-case basis--this is
not a blanket reccomendation for how to address this tool.