Fix issue with spotlight getting wrapped in editor
The technique used to default spotlight to alignfull was causing it to
be wrapped in a div in the editor. Apart from needlessly complicating
the DOM, this also prevented CSS sibling selectors from working, which
is necessary for handling default marings & spacers.

This comment resolves the issue, and should cause no problems with
existing spotlights.