Put get_theme_mod fallbacks in shiro-admin
A common pattern in the templates is to get a string with
`get_theme_mod()` and then provide a fallback with `__('string', 'shiro')`.

I've switched all of these to the 'shiro-admin' namespace because their
current system for localizing these items is to use the Customizer to
set them on each site--bypassing any gettext-based solution. Since one
of our goals with splitting the namespace is to make the 'shiro'
namespace as small and easily-managable as possible I felt it was
prudent to move these items to a different namespace, even if they could
potentially display data on the frontend (not the admin).

All such changes have been collected in this commit so that if we
determine this was a bad choice the change will be easy to revert.