Move all customizer strings to shiro-admin
All these strings have been moved to `shiro-admin` because all
customizer strings are only shown in the admin. Even defaults are only
shown in the admin--not sent to the frontend. Although some (i.e. the
footer) do share defaults with the frontend, client practice is to use
the Customizer to localize these terms, not i18n tools.