Build hm-gutenberg-tools w/ theme
This adds commands to the theme's package.json that run the build/start
commands for the tools plugin along with the equivalent commands for the
the theme build process. This is a little bit of a hack--a more robust
solution would involve creating a unified pipeline for all assets--but
given the already-piecemeal nature of this project, that kind of an
effort should be put off until a later phase.