Limit certain blocks to certain post types
There are a few blocks that don't make sense (or will break) when used
on post types other than the ones they were designed for--the most
obvious (and the inspiration for this change) being the use of "read
more categories" on pages, which don't have categories. This collects a
few of those and limits them to the correct post type.