Add untracked plugins to phpcs_skip file
This tells the VIP Go PHPCS bot to ignore plugins that are not
tracked--that is, plugins that are installed by composer as part of the
build process, not managed directly through git.

The rationale behind this is that we aren't in a position to control the
code of these plugins, so errors, etc, in PRs just amounts to noise and
interferes with catching errors we are in a position to control.

The implication here is that plugins should be audited careful before
being managed by composer.